refactor BudgetMonthCountContext to tsx
